PythonStudy/Smart2/smart/requirements.txt
asgiref==3.7.2
Django==5.0
django-debug-toolbar==4.2.0
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Markdown==3.5.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
